Add logger, refactor name, remove unused params
