+++ app/src/main/java/com/tpp/theperiodpurse/ui/component/LoadingScreen.kt
@@ -1,24 +1,17 @@
-package com.tpp.theperiodpurse.ui.onboarding
+package com.tpp.theperiodpurse.ui.component
 
-import android.os.Build
-import androidx.annotation.RequiresApi
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.size
 import androidx.compose.material.CircularProgressIndicator
-import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.unit.dp
 import com.tpp.theperiodpurse.R
 
-@RequiresApi(Build.VERSION_CODES.O)
 @Composable
 fun LoadingScreen() {
     val redLoading = Color(195, 50, 50)
